refactor(root): change folder structure
