+++ .vscode/extensions.json
@@ -1,0 +1,7 @@
+{
+  "recommendations": [
+    "dsznajder.es7-react-js-snippets",
+    "mikestead.dotenv",
+    "esbenp.prettier-vscode"
+  ]
+}
