+++ package.json
@@ -7,8 +7,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint ./components/** ./pages/** -c .eslintrc.js --ext js,jsx",
-    "lint-fix": "eslint ./components/** ./pages/** -c .eslintrc.js --fix --ext js,jsx"
+    "lint": "eslint ./src/components/** ./src/pages/** -c .eslintrc.js --ext js,jsx",
+    "lint-fix": "eslint ./src/components/** ./src/pages/** -c .eslintrc.js --fix --ext js,jsx"
   },
   "dependencies": {
     "@chakra-ui/react": "^1.8.8",
