+++ src/pages/blog/index.js
@@ -4,8 +4,8 @@ import { NextSeo } from 'next-seo';
 import { createClient } from 'contentful';
 import { Divider, Heading, Text, VStack } from '@chakra-ui/react';
 
-import Card from '../../components/blog/Card';
-import PageContainer from '../../components/PageContainer';
+import Card from '@/components/blog/Card';
+import PageContainer from '@/components/PageContainer';
 
 const client = createClient({
   space: process.env.NEXT_PUBLIC_CONTENTFUL_SPACE_ID,
