+++ src/pages/linktree.js
@@ -2,8 +2,8 @@ import { NextSeo } from 'next-seo';
 import { createClient } from 'contentful';
 import { Divider, Heading, Stack, Text } from '@chakra-ui/react';
 
-import PageContainer from '../components/PageContainer';
-import SocialButton from '../components/about/SocialButton';
+import PageContainer from '@/components/PageContainer';
+import SocialButton from '@/components/about/SocialButton';
 
 const client = createClient({
   space: process.env.NEXT_PUBLIC_CONTENTFUL_SPACE_ID,
